Issues #38 (entities) and #39. Additionally more webflows added. Code formatted to fit Sun checkstyle (no tabs).

+++ SelfDiagnosis/src/main/java/com/selfdiagnosis/web/StringToEntityConverterFactory.java
@@ -8,31 +8,31 @@ import com.selfdiagnosis.core.service.AdminService;
 
 public class StringToEntityConverterFactory implements ConverterFactory<String, SelfDiagnosisEntity>{
 
-	public AdminService adminService;
+    public AdminService adminService;
 
-	public StringToEntityConverterFactory(AdminService adminService) {
-		this.adminService = adminService;
-	}
-	
-	@Override
-	public <T extends SelfDiagnosisEntity> Converter<String, T> getConverter(Class<T> targetType) {
-		return new StringToEntityConverter<T>(targetType);
-	}
-	
-	private final class StringToEntityConverter<T extends SelfDiagnosisEntity> implements Converter<String, T> {
+    public StringToEntityConverterFactory(AdminService adminService) {
+        this.adminService = adminService;
+    }
+    
+    @Override
+    public <T extends SelfDiagnosisEntity> Converter<String, T> getConverter(Class<T> targetType) {
+        return new StringToEntityConverter<T>(targetType);
+    }
+    
+    private final class StringToEntityConverter<T extends SelfDiagnosisEntity> implements Converter<String, T> {
 
-		private Class<T> entityType;
-		
-		public StringToEntityConverter(Class<T> entityType) {
-			this.entityType = entityType;
-		}
-		
-		@Override
-		public T convert(String id) {
-			return adminService.getEntityByTypeAndId(entityType, Integer.valueOf(id));
-		}
-		
-	}
-	
+        private Class<T> entityType;
+        
+        public StringToEntityConverter(Class<T> entityType) {
+            this.entityType = entityType;
+        }
+        
+        @Override
+        public T convert(String id) {
+            return adminService.getEntityByTypeAndId(entityType, Long.valueOf(id));
+        }
+        
+    }
+    
 
 }
